+++ src/Unchase.OData.ConnectedService/ViewModels/ConfigODataEndpointViewModel.cs
@@ -3,13 +3,18 @@
 // Licensed under the MIT License.  See License.txt in the project root for license information.
 
 using System;
+using System.ComponentModel;
 using System.Data.Services.Design;
 using System.Globalization;
 using System.IO;
 using System.Linq;
 using System.Net;
+using System.Net.Security;
 using System.Security;
+using System.Security.Cryptography.X509Certificates;
+using System.Text;
 using System.Threading.Tasks;
+using System.Windows;
 using System.Xml;
 using Microsoft.VisualStudio.ConnectedServices;
 using Unchase.OData.ConnectedService.Common;
@@ -18,13 +23,55 @@ using Unchase.OData.ConnectedService.Views;
 
 namespace Unchase.OData.ConnectedService.ViewModels
 {
-    internal class ConfigODataEndpointViewModel : ConnectedServiceWizardPage
-    {
-        public string Endpoint { get; set; }
+	internal class ConfigODataEndpointViewModel : ConnectedServiceWizardPage
+	{
+        #region Properties and fields
+        private string _endPoint;
+        public string Endpoint
+        {
+            get => _endPoint;
+            set
+            {
+                _endPoint = value;
+                UserSettings.Endpoint = value;
+                OnPropertyChanged(nameof(EndPoint));
+            }
+        }
 
-        public string ServiceName { get; set; }
+		private string _serviceName;
+        public string ServiceName
+        {
+            get => _serviceName;
+            set
+            {
+                _serviceName = value;
+                UserSettings.ServiceName = value;
+                OnPropertyChanged(nameof(ServiceName));
+            }
+        }
 
-        public Version EdmxVersion { get; set; }
+		private Version _edmxVersion;
+        public Version EdmxVersion
+        {
+            get => _edmxVersion;
+            set
+            {
+                _edmxVersion = value;
+                OnPropertyChanged(nameof(EdmxVersion));
+            }
+        }
+
+		private LanguageOption _languageOption;
+        public LanguageOption LanguageOption
+        {
+            get => _languageOption;
+            set
+            {
+                _languageOption = value;
+                UserSettings.LanguageOption = value;
+                OnPropertyChanged(nameof(LanguageOption));
+            }
+        }
 
         public string MetadataTempPath { get; set; }
 
@@ -55,7 +102,9 @@ namespace Unchase.OData.ConnectedService.ViewModels
                     .ToArray();
             }
         }
+        #endregion
 
+        #region Constructors
         public ConfigODataEndpointViewModel(UserSettings userSettings) : base()
         {
             this.Title = "Configure metadata endpoint";
@@ -70,7 +119,9 @@ namespace Unchase.OData.ConnectedService.ViewModels
             this.UseWebProxyCredentials = false;
             this.UserSettings.LanguageOption = LanguageOption.GenerateCSharpCode;
         }
+        #endregion
 
+        #region Methods
         public override Task<PageNavigationResult> OnPageLeavingAsync(WizardLeavingArgs args)
         {
             UserSettings.AddToTopOfMruList(((Wizard)this.Wizard).UserSettings.MruEndpoints, this.Endpoint);
@@ -93,6 +144,7 @@ namespace Unchase.OData.ConnectedService.ViewModels
         }
 
         private string GetMetadata(out Version edmxVersion)
+
         {
             if (string.IsNullOrEmpty(this.UserSettings.Endpoint))
             {
@@ -105,7 +157,7 @@ namespace Unchase.OData.ConnectedService.ViewModels
                     this.UserSettings.Endpoint = this.UserSettings.Endpoint.TrimEnd('/') + "/$metadata";
             }
 
-            var xmlUrlResolver = new XmlUrlResolver()
+            var xmlUrlResolver = new XmlUrlResolver
             {
                 Credentials = this.UseNetworkCredentials ? new NetworkCredential(this.NetworkCredentialsUserName, this.NetworkCredentialsPassword, this.NetworkCredentialsDomain) : CredentialCache.DefaultNetworkCredentials
             };
@@ -132,30 +184,81 @@ namespace Unchase.OData.ConnectedService.ViewModels
 
             try
             {
-                using (var reader = XmlReader.Create(this.UserSettings.Endpoint, readerSettings))
-                {
-                    using (var writer = XmlWriter.Create(workFile))
-                    {
-                        while (reader.NodeType != XmlNodeType.Element)
-                        {
-                            reader.Read();
-                        }
-
-                        if (reader.EOF)
-                        {
-                            throw new InvalidOperationException("The metadata is an empty file");
-                        }
-
-                        Common.Constants.SupportedEdmxNamespaces.TryGetValue(reader.NamespaceURI, out edmxVersion);
-                        writer.WriteNode(reader, false);
-                    }
-                }
-                return workFile;
+                return ReadMetadata(out edmxVersion, readerSettings, workFile);
             }
             catch (WebException e)
             {
+                if (e.InnerException is System.Security.Authentication.AuthenticationException)
+                {
+                    var save = ServicePointManager.ServerCertificateValidationCallback;
+                    ServicePointManager.ServerCertificateValidationCallback = LifeValidationCallback;
+                    return ReadMetadata(out edmxVersion, readerSettings, workFile);
+                }
                 throw new InvalidOperationException(string.Format(CultureInfo.InvariantCulture, "Cannot access {0}", this.UserSettings.Endpoint), e);
             }
         }
+
+        /// <summary>
+        /// Read the metadata.
+        /// </summary>
+        /// <param name="edmxVersion"><see cref="Version"/> of edmx.</param>
+        /// <param name="readerSettings"><see cref="XmlReaderSettings"/>.</param>
+        /// <param name="workFile"></param>
+        /// <returns></returns>
+        private string ReadMetadata(out Version edmxVersion, XmlReaderSettings readerSettings, string workFile)
+        {
+            using (var reader = XmlReader.Create(this.UserSettings.Endpoint, readerSettings))
+            {
+                using (var writer = XmlWriter.Create(workFile))
+                {
+                    while (reader.NodeType != XmlNodeType.Element)
+                    {
+                        reader.Read();
+                    }
+
+                    if (reader.EOF)
+                    {
+                        throw new InvalidOperationException("The metadata is an empty file");
+                    }
+
+                    Constants.SupportedEdmxNamespaces.TryGetValue(reader.NamespaceURI, out edmxVersion);
+                    writer.WriteNode(reader, false);
+                }
+            }
+            return workFile;
+        }
+
+        /// <summary>
+        /// Validation of possibly untrusted certificates.
+        /// </summary>
+        /// <param name="sender"></param>
+        /// <param name="certificate"><see cref="X509Certificate"/>.</param>
+        /// <param name="chain"><see cref="X509Chain"/>.</param>
+        /// <param name="sslPolicyErrors"><see cref="SslPolicyErrors"/>.</param>
+        /// <returns></returns>
+        private bool LifeValidationCallback(object sender, X509Certificate certificate, X509Chain chain, SslPolicyErrors sslPolicyErrors)
+        {
+            var sb = new StringBuilder();
+
+            var x509 = certificate as X509Certificate2;
+            sb.AppendLine();
+            sb.AppendLine($"Content Type: {X509Certificate2.GetCertContentType(x509?.RawData ?? new byte[]{})}");
+            sb.AppendLine($" Name: {certificate.Subject}");
+            sb.AppendLine($"Certificate Verified?: {x509?.Verify()}");
+            sb.AppendLine($"Simple Name: {x509?.GetNameInfo(X509NameType.SimpleName, true)}");
+            sb.AppendLine($"Signature Algorithm: {x509?.SignatureAlgorithm.FriendlyName}");
+            //sb.AppendLine($"Private Key: {x509?.PrivateKey.ToXmlString(false)}");
+            //sb.AppendLine($"Public Key: {x509?.PublicKey.Key.ToXmlString(false)}");
+            sb.AppendLine($"Certificate Archived?: {x509?.Archived}");
+            sb.AppendLine($"Length of Raw Data: {x509?.RawData.Length}");
+
+            var result = MessageBox.Show($"Would you like to Accept the untrusted Certificate:{Environment.NewLine}{sb}", "Attention", MessageBoxButton.YesNo, MessageBoxImage.Question);
+            if (result != MessageBoxResult.Yes)
+                return false;
+
+            return true;
+        }
+        #endregion
     }
 }
+